link module JS file
