+++ index.html
@@ -249,6 +249,6 @@
   </template>
 
 </body>
-<script src="js/main.js"></script>
+<script src="/js/main.js" type="module"></script>
 
 </html>
